api/common-utils/src/main/java/com/roche/connect/adm/dto/TokenDetails.java
package com.roche.connect.adm.dto;

import java.util.Date;

public class TokenDetails {
	
	private String userName;
	private String domainName;
	private Date issuedDate;
	private Date expiryDate;
	public String getUserName() {
		return userName;
	}
	public void setUserName(String userName) {
		this.userName = userName;
	}
	public String getDomainName() {
		return domainName;
	}
	public void setDomainName(String domainName) {
		this.domainName = domainName;
	}
	public Date getIssuedDate() {
		return issuedDate;
	}
	public void setIssuedDate(Date issuedDate) {
		this.issuedDate = issuedDate;
	}
	public Date getExpiryDate() {
		return expiryDate;
	}
	public void setExpiryDate(Date expiryDate) {
		this.expiryDate = expiryDate;
	}
	
	

}
